- Baked the terrain - please don't change  - Polished the credits menu  - Polished the end screen  - Fixed a few bugs relating to the end state  -

+++ Roller Derby/Assets/Scripts/GameController.cs
@@ -35,7 +35,7 @@ public class GameController : MonoBehaviour
     {
         currentTime -= Time.deltaTime;
 
-        timer.text = currentTime.ToString();
+        timer.text = ((int)currentTime).ToString();
 
         if (currentTime <= 0)
             LoseGame();
